Chore: Add services

+++ aplicação/backend/src/main/java/eaproject/input/UpdateLocationInput.java
@@ -6,8 +6,8 @@ import java.sql.Timestamp;
 public class UpdateLocationInput extends BaseInput implements Serializable {
 
     private int id;
-    private float latitude;
-    private float longitude;
+    private Double latitude;
+    private Double longitude;
     private String address;
     private String city;
     private String country;
@@ -22,19 +22,19 @@ public class UpdateLocationInput extends BaseInput implements Serializable {
         this.id = id;
     }
 
-    public float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
