+++ aplicação/backend/src/main/java/eaproject/output/GetLocationByIdOutput.java
@@ -11,10 +11,10 @@ public class GetLocationByIdOutput extends BaseOutput implements Serializable {
     private int id;
 
     @JsonProperty
-    private float latitude;
+    private Double latitude;
 
     @JsonProperty
-    private float longitude;
+    private Double longitude;
 
     @JsonProperty
     private String address;
@@ -39,19 +39,19 @@ public class GetLocationByIdOutput extends BaseOutput implements Serializable {
         this.id = id;
     }
 
-    public float getLatitude() {
+    public Double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(Double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public Double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(Double longitude) {
         this.longitude = longitude;
     }
